+++ src/main/java/com/okestro/community/service/BoardService.java
@@ -8,6 +8,6 @@ public interface BoardService {
 
     List<ListDto> boardlist();
 
-    ListDto getBoardById(int id);
+    ListDto getBoardById(Long id);
 
 }
